Move STREAMS python helpers under scripts
