+++ README
@@ -561,7 +561,7 @@ module combinations and print the average, minimum and maximum latency
 for each:
 
 ```
-$ python3 scripts/simulate.py
+$ python3 scripts/streams_latency.py
 xor: avg 0.50 us (min 0.40, max 0.60)
 upper+xor: avg 0.70 us (min 0.60, max 0.80)
 reverse+upper+xor: avg 0.90 us (min 0.80, max 1.00)
@@ -582,7 +582,7 @@ the numeric values::
 
 Missing or malformed files fall back to built-in defaults. Writing a
 new value to one of these files immediately updates the active
-constant. The ``flow_pid.py`` helpers ``set_kp()``, ``set_ki()`` and
+constant. The ``scripts/flow_pid.py`` helpers ``set_kp()``, ``set_ki()`` and
 ``set_kd()`` provide a convenient API for adjusting the parameters.
 
 For testing the base directory can be overridden with the environment
@@ -593,14 +593,14 @@ with a Python example is provided in
 STREAMS LOGGING
 ---------------
 Python STREAMS modules log messages using ``strlog_json()`` from
-``streams_log.py``.  The helper outputs a JSON object containing the
+``scripts/streams_log.py``.  The helper outputs a JSON object containing the
 timestamp, log level and message.  Extra keyword arguments become
 fields in the record.  ``strlog()`` continues to print plain text for
 older modules.
 
 Example::
 
-    from streams_log import strlog_json
+    from scripts.streams_log import strlog_json
     strlog_json("info", "module loaded", name="xor_encrypt")
 
 AUTOMATED SETUP WITH CODEX
