+++ TODO_STREAMS.md
@@ -16,12 +16,12 @@ This file collects outstanding tasks for the prototype STREAMS implementation. T
 
 ## Testing and tooling
 
-- Expand the latency harness in `scripts/simulate.py` to cover more module combinations.
-- Add unit tests for the logging helpers in `streams_log.py`.
+- Expand the latency harness in `scripts/streams_latency.py` to cover more module combinations.
+- Add unit tests for the logging helpers in `scripts/streams_log.py`.
 
 ## Flow control and configuration
 
 - Document the PID based flow control interface under `/proc/streams/fc/`.
-- Provide an example illustrating dynamic tuning via `flow_pid.py`.
+- Provide an example illustrating dynamic tuning via `scripts/flow_pid.py`.
 
 Additional issues may be recorded here as the prototype evolves.
