+++ tests/test_encrypt_mod.py
@@ -4,8 +4,8 @@ import sys
 ROOT = pathlib.Path(__file__).resolve().parents[1]
 sys.path.insert(0, str(ROOT))
 
-from encrypt_mod import XorEncryptModule
-from stream_module import mblk_t
+from scripts.encrypt_mod import XorEncryptModule
+from scripts.stream_module import mblk_t
 
 
 def test_xor_encrypt_decrypt_roundtrip():
